requirements.txt
Flask==2.3.3
requests==2.31.0
g4f==0.1.4.1
python-dotenv==1.0.0